refactor: 반복되는 공통 타입, 인터페이스를 외부로 분리
